Move strategies to new package
